Generate supported tags
